changed strong monthly reward
changed strong monthly reward to 2.821 was set at 2.76 which was a 30 day projection. all month are set to 31 days as theres more month with 31 days than anything else. in future would like to have date selector, or start from current data. possibly even save a projection and load it on the fly.